% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{enricher}
\alias{enricher}
\title{Pathway Enrichment Filtering and Plotting}
\usage{
enricher(data = NULL, clusters = NULL, genes = NULL, pathways = NULL)
}
\arguments{
\item{data}{Clustered results from pathfindR experiment using pathfindR cluster_enriched_terms(data, plot_dend = TRUE, plot_clusters_graph = TRUE)}

\item{clusters}{A numeric vector of cluster IDs to filter the data. Default is NULL.}

\item{genes}{A vector of gene names to filter the data. Default is NULL.}

\item{pathways}{A vector of pathway names to filter the data. Default is NULL.}
}
\value{
A plot object created with the filtered data by the pathfindR enrichment_chart() method
}
\description{
This function allows you to filter a dataset of pathway enrichment data by specific clusters, genes, or pathways and visualize the resulting enrichment using a plot.
}
\details{
This function filters a dataset based on provided cluster #'s, genes, or pathways and generates an enrichment chart.
}
\examples{
# Example usage:
# enricher(data, clusters = c(1), genes = c("Gene1"), pathways = c("Pathway A"))
}
