% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{erupt}
\alias{erupt}
\title{Create a Volcano Plot for Gene Expression}
\usage{
erupt(
  data = NULL,
  down_reg = -1,
  up_reg = 1,
  pval = 0.05,
  population = 0.3,
  lab_density = 0.25,
  genes = c("NOTHING TO SEARCH"),
  title = "",
  subtitle = "",
  caption = ""
)
}
\arguments{
\item{data}{A data frame containing gene expression data.}

\item{down_reg}{Numeric, threshold for down-regulation (default: -1).}

\item{up_reg}{Numeric, threshold for up-regulation (default: 1).}

\item{pval}{Numeric, threshold for p-value significance (default: 0.05).}

\item{population}{Numeric, proportion of the population to consider (default: 0.3).}

\item{lab_density}{Numeric, density of labels to be displayed (default: 0.25).}

\item{genes}{Character vector, specific genes to search and highlight (default: "NOTHING TO SEARCH").}

\item{title}{Character, title of the plot (default: "").}

\item{subtitle}{Character, subtitle of the plot (default: "").}

\item{caption}{Character, caption of the plot (default: "").}
}
\value{
A volcano plot object.
}
\description{
This function generates a volcano plot to visualize gene expression data.
}
