% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_heatmap_gene.R
\name{geneheatmap}
\alias{geneheatmap}
\title{This function filters a dataset based on provided genes and returns plots}
\usage{
geneheatmap(data = NULL, genes = NULL)
}
\arguments{
\item{data}{Clustered results from pathfindR experiment using pathfindR cluster_enriched_terms(data, plot_dend = TRUE, plot_clusters_graph = TRUE)}

\item{genes}{A vector of gene names to filter the data. Default is NULL.}
}
\value{
A list (ggplot, plotly) displaying the data
}
\description{
This function allows you to filter a dataset of pathway enrichment data by a character vector of genes, a ggplot and plotly are both returned respectivly in a list.
}
\examples{
# Example usage:
# geneheatmap(data, genes = c("Gene1", "Gene2"))
}
