% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_heatmap_pathway.R
\name{pathwayheatmap}
\alias{pathwayheatmap}
\title{This function filters a dataset based on provided pathways and returns plots}
\usage{
pathwayheatmap(data = NULL, pathways = NULL)
}
\arguments{
\item{data}{Clustered results from pathfindR experiment using pathfindR cluster_enriched_terms(data, plot_dend = TRUE, plot_clusters_graph = TRUE)}

\item{pathways}{Either a numeric that determines the number of genes to visualize or a char vector of gene names to visualize. Default is NULL.}
}
\value{
A list (ggplot, plotly) displaying the data
}
\description{
This function allows you to filter a dataset of pathway enrichment data by a character vector of pathways or a desired number of pathways to visuallize, a ggplot and plotly are both returned respectively in a list.
}
\examples{
# Example usage:
# pathwayheatmap(data, pathways = c("pathway1", "pathway2"))
# pathwayheatmap(data, pathways = 20)
}
