% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_terms_plot.R
\name{score_pathway_terms}
\alias{score_pathway_terms}
\title{This function filters a dataset based on provided pathways}
\usage{
score_pathway_terms(
  data = NULL,
  abundance = NULL,
  cases = NULL,
  repOnly = TRUE,
  pathways = NULL
)
}
\arguments{
\item{data}{Clustered results from pathfindR experiment using pathfindR cluster_enriched_terms(data, plot_dend = TRUE, plot_clusters_graph = TRUE)}

\item{abundance}{a dataframe of normalized counts must contain 1 column of Gene_symbol}

\item{cases}{A vector of sample names that belong to the experiment case}

\item{repOnly}{True if to only plot the representative status pathways}

\item{pathways}{A vector of pathway names to filter the data. Default is NULL.}
}
\value{
A list (ggplot, height, width) displaying the data
}
\description{
This function allows you to filter a dataset of pathway enrichment data by a character vector of pathways, a ggplot and recommended dimensions are returned in a list.
}
\examples{
# Example usage:
# score_pathways(data, abundance, cases, repOnly = TRUE, pathways = c('pathway1'))
}
