% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pathway_sample_map.R
\name{single_pathway_heatmap}
\alias{single_pathway_heatmap}
\title{Create a heatmap for Gene Expression}
\usage{
single_pathway_heatmap(
  pathway = NULL,
  DEG_results = NULL,
  Pathfinder_results = NULL,
  genes_listed = 100
)
}
\arguments{
\item{DEG_results}{A data frame containing gene expression data.}

\item{Pathfinder_results}{A data frame containing pathway expression data}

\item{genes_listed}{the number of pathway genes to include ranked by padj}

\item{pathwya}{a sting of the searched pathway to match}
}
\value{
A ggplot heatmap
}
\description{
Create a heatmap for Gene Expression
}
